Moved smhm_components to its own subpackage
