+++ halotools/empirical_models/__init__.py
@@ -7,7 +7,6 @@ from .mock_factories import *
 from .model_factories import *
 from .preloaded_models import *
 from .model_defaults import *
-from .smhm_components import *
 from .abunmatch import *
 from .model_helpers import *
 from .assembias_models import *
@@ -15,3 +14,4 @@ from .assembias_models import *
 from .phase_space_models import *
 from .sfr_models import *
 from .occupation_models import *
+from .smhm_models import *
