+++ halotools/empirical_models/occupation_models/hod_components.py
@@ -20,7 +20,8 @@ from astropy.extern import six
 from abc import ABCMeta, abstractmethod, abstractproperty
 import warnings
 
-from .. import model_defaults, model_helpers, smhm_components
+from .. import model_defaults, model_helpers
+from ..smhm_models import smhm_components
 from ..assembias_models import HeavisideAssembias
 from ..model_helpers import bounds_enforcing_decorator_factory
