+++ halotools/empirical_models/occupation_models/tinker13_components.py
@@ -22,7 +22,9 @@ import warnings
 
 from .hod_components import OccupationComponent 
 
-from .. import model_defaults, model_helpers, smhm_components
+from .. import model_defaults, model_helpers
+from ..smhm_models import smhm_components
+
 from ..assembias_models import HeavisideAssembias
 from ..model_helpers import bounds_enforcing_decorator_factory
