+++ halotools/empirical_models/occupation_models/zheng07_components.py
@@ -23,7 +23,8 @@ import warnings
 
 from .hod_components import OccupationComponent
 
-from .. import model_defaults, model_helpers, smhm_components
+from .. import model_defaults, model_helpers
+from ..smhm_models import smhm_components
 from ..assembias_models import HeavisideAssembias
 from ..model_helpers import bounds_enforcing_decorator_factory
