+++ halotools/empirical_models/smhm_models/__init__.py
@@ -1,0 +1,1 @@
+from smhm_components import *
